Restore absolute line numbers in the background after a fast open
Seeking straight to the tail on open (previous commit) means the
gutter numbers what's in the buffer, not the file - a 5-million-line
file would show "1" at the top of the window instead of its real line
number. That's a legitimate tradeoff for keeping opens fast, but a bare
number with no context just looks wrong rather than intentional.

wc -l is fast for the same reason FileObserverImpl's read fix works:
scanning bytes linearly and counting '\n' is cheap; it's the
repeated re-reading that was slow. LineCounter applies the same trick
to count the lines skipped by the tail-seek in the background, off the
EDT, and RingLineBuffer.applyLineNumberOffset() applies the correction
once it's done - additively, so it composes correctly with any
eviction that already happened while the count was running. In
practice this finishes fast enough that real numbers are usually
already in place by the first repaint.

firstLineNumber widens from int to long throughout LineBuffer, since a
tool now explicitly built to handle huge files shouldn't silently wrap
around on one big enough to need it.

+++ src/main/java/com/selesse/tailerswift/gui/lines/LineNumberGutter.java
@@ -57,7 +57,7 @@ public class LineNumberGutter extends JComponent {
         if (fontMetrics == null) {
             return;
         }
-        int highestLineNumber = lineBuffer.getFirstLineNumber() + lineBuffer.size();
+        long highestLineNumber = lineBuffer.getFirstLineNumber() + lineBuffer.size();
         int digits = Math.max(String.valueOf(highestLineNumber).length(), MIN_DISPLAY_DIGITS);
 
         if (digits != lastDigits) {
